==> Code/0_Downloading Tweets to a CSV.ipynb <==
Code for extracting tweets from the twitter API, mostly taken from Grace's code

#import necessary packages
import csv
import tweepy

# add the access codes from your account
access_key=''
access_secret=''
consumer_key=''
consumer_secret=''

auth = tweepy.OAuthHandler(consumer_key, consumer_secret)
auth.set_access_token(access_key, access_secret)
api = tweepy.API(auth, wait_on_rate_limit=True)

csvFile = open('vaccine_college_500_2.csv', 'a') # create & title your csv file
csvWriter = csv.writer(csvFile)
column_headers=['date', 'text', 'retweet_count', 'favorite_count', 'reply']

search_words='vaccine_college' #put in your preferred keywords
new_search = search_words + " -filter:retweets" #filter out retweets

csvWriter.writerow(column_headers)
for tweet in tweepy.Cursor(api.search, q=new_search, count=500,
                        lang='en',
                        since_id=0, 
                        tweet_mode='extended').items():
    csvWriter.writerow([tweet.created_at,
                        tweet.full_text.encode('ascii', 'ignore'),
                        tweet.retweet_count,
                        tweet.favorite_count,
                        1 if tweet.in_reply_to_status_id
                        else 0])
