==> Code/1_vader_r.Rmd <==
---
title: "Final Dataset Vader"
author: "Jakob Woerner"
date: "`r Sys.Date()`"
output: 
  html_document:
    theme: 'united'
    highlight: 'tango'
    toc: yes
    toc_float: true
    code_folding: 'show'
    df_print: 'paged'
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,comment = NA)
```

# Setup
```{r, warning=FALSE, message=FALSE}
if("vader" %in% rownames(installed.packages()) == FALSE) {install.packages("vader")}
library(vader);library(xlsx);library(tidyverse)
```

# Goal

Run cleaning and then Vader sentiment analysis on the final dataset


# Load Your Here Data

```{r, warning=TRUE, message=FALSE}
# put the tweet csv file path here
tweet_raw <- read_csv("../Data/all_tweet_data.csv")

```

# Clean Data

Remove first 'b'

```{r}
vc1 <- tweet_raw %>% 
  mutate(clean_text = str_sub(text, 2)) %>% 
  select(clean_text, everything())
```

Remove URLs and newline characters

```{r}
vc2 <- vc1 %>% 
  mutate(clean_text = gsub("http.*","",  clean_text)) %>% 
  mutate(clean_text = gsub("https.*","",  clean_text)) %>% 
  mutate(clean_text = gsub("\\\\n","",  clean_text))
```

Let's not make lowercase or remove special characters, since those might be used in the vader algorithm.

Remove @, quotation marks, and apostrophes

```{r}
vc2$clean_text <- gsub("'", '', vc2$clean_text)
vc2$clean_text <- gsub('"', '', vc2$clean_text)
vc2$clean_text <- gsub("@\\w+", "", vc2$clean_text)
```

Remove duplicates and pincode tweets

```{r}
vc3 <- vc2[!duplicated(vc2$clean_text),] %>% 
  filter(!str_detect(text, "Pin Code|Pincode"))
```

# Apply vader    

```{r}
# this might take a while to run
#t1 <- Sys.time()
vader_result <- vader_df(vc3$clean_text)
#t2 <- t1 - Sys.time()
#t2-t1
```

There might be a problematic tweet which caused an error/warning

"number of items to replace is not a multiple of replacement length1 row contains an error. Filter word_scores for 'ERROR' to identify the problematic text.Incompatible methods ("Ops.difftime", "-.POSIXt") for "-""

# Neutralize

Recode to remove the neutral category.

```{r}
vader_no_neutral <- vader_result %>% 
  drop_na() %>% # 6 rows dropped
  mutate(positive = pos / (pos + neg),
         negative = neg / (pos + neg),
         neutral = if_else(neu == 1, 1, 0),
         assigned = if_else(neu == 1, "Neutral",
                            if_else(positive > 0.5, "Positive","Negative")))
```

# Save data

```{r}
vader_final <- vader_no_neutral %>% 
  select(text, positive, negative, assigned) %>% 
  drop_na()

write.csv(vader_final, "../Data/vader_result_final.csv")
```


# Session Information
```{r}
sessioninfo::session_info()
```
